+++ CHANGELOG.md
@@ -1,3 +1,106 @@
+Drupal Cookbook CHANGELOG
+======================
+This file is used to list changes made in each version of the drupal cookbook.
+
+## v1.1.21
+------
+### Improvement
+- Allow force install
+
+## v1.1.20
+------
+### Improvement
+- Support Drupal 6
+- Support stripping comments out of settings.php
+
+## v1.1.19
+------
+### Improvement
+- Notify mysql grants creation
+
+## v1.1.18
+------
+### Bug
+- Prevent site-install from being triggered repeatedly!
+
+## v1.1.17
+------
+### Improvement
+- Refuse to (re)install the site when the database host is not 127.0.0.1. We do NOT want to wipe out content!
+
+## v1.1.16
+------
+### Bug
+- Fix dependencies of sites/default/files/settings.php
+
+## v1.1.15
+------
+### Bug
+- Fix fatal log for sites/default/settings.php from sites/default/files/settings.php
+
+## v1.1.14
+------
+### Improvement
+- Allow template to be created and maintained without triggering repeated site installs!
+
+## v1.1.13
+------
+### Improvement
+- Modules will not install until the site-install is complete and was successful
+
+## v1.1.12
+------
+### Improvement
+- Creating the settings.php based on Drupal unpack chain has pitfalls. Let's just create it if it is missing.
+
+## v1.1.11
+------
+### Bug
+- Set rx on all directories of the DocumentRoot path
+
+## v1.1.10:
+
+* Finetune the Drupal docroot file and directory permissions
+
+## v1.1.9:
+
+* Standardize the default[:drupal][:db] array
+
+## v1.1.8:
+
+* Set Drupal docroot permissions correctly
+* Allow web_app to be disabled
+
+## v1.1.7:
+
+* Correct hostsfile_entry name
+* ctools module installed by default because of webform
+* patches support
+
+## v1.1.6:
+
+* hostsfile to manage /etc/hosts
+
+## v1.1.5:
+
+* Rework dependencies and sequence to support non-root credentials for mysql
+* Set permissions on docroot
+* Restart Apache if necessary
+
+## v1.1.4:
+
+* Allow us to use another drush cookbook
+
+## v1.1.3:
+
+* Don't install drush by default
+
+## v1.1.2:
+
+* Default to 7.23
+* Consider 127.0.0.1 localhost also
+* 
+
 ## v1.1.0:
 
 * fix bug caused by the removal of `mysql_database` LWRP from the mysql cookbook starting with version 1.2.0